Inserido validação para os métodos de create and delete

+++ api/src/core/application/use-cases/custumer/commands/delete-customer.use-command.ts
@@ -1,6 +1,8 @@
 import { CustomerRepository } from "src/core/infrastructure/Repositories/customer.repository"
 import { CommandHandler, ICommandHandler } from "@nestjs/cqrs"
 import { Result, result } from "src/core/infrastructure/Shared/result.util";
+import { messages } from "src/core/infrastructure/Shared/messages";
+import { BadRequestException } from "@nestjs/common";
 
 export class DeleteCustomerCommand {
   constructor(public readonly id: number) {}
@@ -13,6 +15,15 @@ export class DeleteCustomerHandler implements ICommandHandler<DeleteCustomerComm
   }
 
   async execute(command: DeleteCustomerCommand): Promise<Result<{ id: number }>> {
+
+
+    const registerExists = await this.repository.getFirstByParameters({
+      id: command.id,
+    });
+
+    if (!registerExists)
+      throw new BadRequestException(messages.CUSTOMER_NOT_FOUND(command.id));
+
     const deletedId = await this.repository.delete(command.id);
     return result({ id: deletedId }).Success();
   }
